fix index && readme

+++ libft/ft_lstsize.c
@@ -6,15 +6,15 @@
 /*   By: mdourdoi <mdourdoi@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/10 13:49:57 by mdourdoi          #+#    #+#             */
-/*   Updated: 2025/11/10 13:54:53 by mdourdoi         ###   ########.fr       */
+/*   Updated: 2026/01/06 17:50:36 by mdourdoi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_lstsize(t_list *lst)
+long	ft_lstsize(t_list *lst)
 {
-	int	res;
+	long	res;
 
 	res = 0;
 	while (lst)
